Created and finished assignment for week 3
I created a week-3 folder and the page.js, item.js, and item-list.js  files inside of it. I then built the item and item-list components. After that, I built the page component using the item-list component and tailwind css styling to display the list of items. Lastly, I updated the metadata's title property to display "Shopping List" and updated the discription to display "Shopping list for groceries".