Finished the week 6 assignment
Created 3 buttons in the item-list file that sorts the shopping list by name, category, or grouped category. Also, instead of hard coding each item into the list, I used a .json file and imported it into item-list and called them using mapping and the id as the key for each different item.