completed week 7 assignment
Created and completed the week 7 assignment using the requirements provided. This time, we use a state variable (items) initialized with the data from the json file. I also changed the functions that mutate the array of items directly so that they make a copy of that array and then modifies that array instead of the original array.  The alert functionality was replaced with the onAddItem prop and a handleSubmit function that creates the new item based on the user inputs when the form is submitted. This summarizes the week 7 assignment.